requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
future==1.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
MarkupSafe==3.0.2
python-dotenv==1.1.0
pywebostv==0.8.9
requests==2.32.3
rpi-lgpio==0.6
rpi-rf==0.9.7
RPi.GPIO==0.7.1
urllib3==2.4.0
wakeonlan==3.1.0
Werkzeug==3.1.3
ws4py==0.6.0
yt-dlp==2025.3.31
zipp==3.21.0